cast data to python types for json

+++ nlisim/modules/antitnfa.py
@@ -93,7 +93,7 @@ class AntiTNFa(MoleculeModel):
         voxel_volume = state.voxel_volume
 
         return {
-            'concentration': np.mean(anti_tnf_a.grid) / voxel_volume,
+            'concentration': float(np.mean(anti_tnf_a.grid) / voxel_volume),
         }
 
     def visualization_data(self, state: State) -> Tuple[str, Optional[Union[CellData, np.ndarray]]]:
