+++ nlisim/modules/hemolysin.py
@@ -78,7 +78,7 @@ class Hemolysin(MoleculeModel):
         voxel_volume = state.voxel_volume
 
         return {
-            'concentration': np.mean(hemolysin.grid) / voxel_volume,
+            'concentration': float(np.mean(hemolysin.grid) / voxel_volume),
         }
 
     def visualization_data(self, state: State) -> Tuple[str, Optional[Union[CellData, np.ndarray]]]:
