+++ nlisim/modules/hepcidin.py
@@ -76,7 +76,7 @@ class Hepcidin(MoleculeModel):
         voxel_volume = state.voxel_volume
 
         return {
-            'concentration': np.mean(hepcidin.grid) / voxel_volume,
+            'concentration': float(np.mean(hepcidin.grid) / voxel_volume),
         }
 
     def visualization_data(self, state: State) -> Tuple[str, Optional[Union[CellData, np.ndarray]]]:
