+++ nlisim/modules/il8.py
@@ -107,7 +107,7 @@ class IL8(MoleculeModel):
         voxel_volume = state.voxel_volume
 
         return {
-            'concentration': np.mean(il8.grid) / voxel_volume,
+            'concentration': float(np.mean(il8.grid) / voxel_volume),
         }
 
     def visualization_data(self, state: State) -> Tuple[str, Optional[Union[CellData, np.ndarray]]]:
