+++ nlisim/modules/iron.py
@@ -69,7 +69,7 @@ class Iron(MoleculeModel):
         voxel_volume = state.voxel_volume
 
         return {
-            'concentration': np.mean(iron.grid) / voxel_volume,
+            'concentration': float(np.mean(iron.grid) / voxel_volume),
         }
 
     def visualization_data(self, state: State) -> Tuple[str, Optional[Union[CellData, np.ndarray]]]:
