+++ nlisim/modules/mip1b.py
@@ -102,7 +102,7 @@ class MIP1B(MoleculeModel):
         voxel_volume = state.voxel_volume
 
         return {
-            'concentration': np.mean(mip1b.grid) / voxel_volume,
+            'concentration': float(np.mean(mip1b.grid) / voxel_volume),
         }
 
     def visualization_data(self, state: State) -> Tuple[str, Optional[Union[CellData, np.ndarray]]]:
