+++ nlisim/modules/ros.py
@@ -62,7 +62,7 @@ class ROS(MoleculeModel):
         voxel_volume = state.voxel_volume
 
         return {
-            'concentration': np.mean(ros.grid) / voxel_volume,
+            'concentration': float(np.mean(ros.grid) / voxel_volume),
         }
 
     def visualization_data(self, state: State) -> Tuple[str, Optional[Union[CellData, np.ndarray]]]:
