+++ nlisim/modules/tnfa.py
@@ -139,7 +139,7 @@ class TNFa(MoleculeModel):
         voxel_volume = state.voxel_volume
 
         return {
-            'concentration': np.mean(tnfa.grid) / voxel_volume,
+            'concentration': float(np.mean(tnfa.grid) / voxel_volume),
         }
 
     def visualization_data(self, state: State) -> Tuple[str, Optional[Union[CellData, np.ndarray]]]:
